pyproject.toml
[project]
authors = [
    {name = "mefamex", website = "https://mefamex.com/"}]
name = "image-meta-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = "MIT"
license-files = ["LICENCE"]
requires-python = ">=3.13"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
